Make Lists Iterable, Add Array Like Methods
The work in this commit:

refactors the LinkedList constructor to accept an iterable instead of head and tail Nodes
makes LinkedLists iterable via iterable iterator interface
renames insertAtIndex to insert and refactors
renames method insertAtHead to addFirst
renames method insertAtTail to addLast
renames method emptyList to clear
refactors addFirst and addLast to be a curry of insert
renames iterateOverList to forEach
renames findAtIndex to at
renames removeAtIndex to remove
adds method filter, toArray, slice, includes
improves list typing and adds typeguard isLinkWithData
adds relevant tests to prevent regression and outline intended behavior
adds dependency @types/sinon
updates README to reflect API changes
adds tsconfig.json

By modelling methods after methods from the Array prototype, I find the method names and
functionality to be clearer and simpler.

The outstanding differences between the function signatures of forEach and filter
compared to their Array counterparts is that an additional argument for finalIndex is
needed to prevent a stack overflow since they would keep cycling. Additionally, These
methods do not accept an argument for binding the execution context `this`

The method name 'at' might be controversial but in my opinion this change is preferable for
a couple of reasons: This method acts as a zero based index so a short method name doesn't
get in the way of chaining method calls or accessing properties and the previous naming of
'findAtIndex' occludes the utility of accepting a callback for execution on the target node

The iteration logic in the method `at` has been delegated to a local function, since a
slice method is now available there is no reason to expose the `previousNode` and
`previousIndex` variables and I am now of the opinion that doing so would be a bad idea.