Rename and refactor methods, add Array-Like methods
renames insertAtIndex to insert and refactors
renames method insertAtHead to addFirst
renames method insertAtTail to addLast
delegates addFirst and addLast by currying insert
renames iterateOverList to forEach and refactors
renames findAtIndex to at and refactors
renames removeAtIndex to remove
adds method filter
adds method toArray
adds method slice
adds method includes
simplifies method clear
introduces types IterationCallback and IterationPredicate
removes BDD tests where they are not needed

addFirst and addLast are now curries of insert, this removes duplication of logic needed
to handle edge cases. One of the biggest take aways from this work is that the insertion
logic (including during instantiation) is much simpler when you think in terms of only
adding one node at a time. Regardless of whether a node is added to the front, or the
back, the only edge case that needs to be considered is when we are adding a node to an
empty list: either there are nodes that exist and need to be linked - or - they aren't.
Where they are is largely irrelevant as long as we update the reference to the head and
tail when it is necessary

By modelling methods after methods from the Array prototype, I find the method names and
functionality to be clearer and simpler.

The outstanding differences between the function signatures of forEach and filter
compared to their Array counterparts is that an additional argument for finalIndex is
needed to prevent a stack overflow since they would keep cycling. Additionally, These
methods do not accept an argument for binding the execution context `this`

The method name 'at' might be controversial but in my opinion this change is preferable for
a couple of reasons: This method acts as a zero based index so a short method name doesn't
get in the way of chaining method calls or accessing properties and the previous naming of
'findAtIndex' occludes the utility of accepting a callback for execution on the target node

The iteration logic in the method `at` has been delegated to a local function, since a
slice method is now available there is no reason to expose the `previousNode` and
`previousIndex` variables and I am now of the opinion that doing so would be a bad idea.